// generated/src/app/api-models/vbi-check-register.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class VbiCheckRegister {
  bankAccountCode: string;
  checkNumber: string;
  checkDate: Date;
  seqClaimId: string;
  seqVendId: number;
  seqVendAddress: number;
  companyCode: string;
  checkAmt: number;
  clearDate: Date;
  checkStatus: string;
  comments: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  seqCkprtId: number;
  eftTransNumber: string;
  manualCheckFlag: string;
  statChngDate: Date;
  statChngRsnCode: string;
  crossReference: string;
  category: string;
  fileType: string;
}